Downloader Class
Moved the the the majority of the program to the "Downloader.py" class, which has simplified the main.py python file.
